refactor: 도메인객체에 경로찾기 라이브러리가 직접적으로 의존하지 않도록 리팩토링
