+++ src/main/java/wooteco/subway/domain/discountstrategy/AgeDiscountStrategyFactory.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.discountstrategy;
 
 public class AgeDiscountStrategyFactory {
