+++ src/main/java/wooteco/subway/domain/discountstrategy/ChildrenDiscountStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.discountstrategy;
 
 public class ChildrenDiscountStrategy implements AgeDiscountStrategy {
     @Override
