+++ src/main/java/wooteco/subway/domain/discountstrategy/TeenagerDiscountStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.discountstrategy;
 
 public class TeenagerDiscountStrategy implements AgeDiscountStrategy {
     @Override
