+++ src/main/java/wooteco/subway/domain/path/PathFinder.java
@@ -1,0 +1,5 @@
+package wooteco.subway.domain.path;
+
+public interface PathFinder {
+    PathResult getPath(Long source, Long target, int age);
+}
