+++ src/main/java/wooteco/subway/domain/path/PathResult.java
@@ -1,6 +1,8 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 import java.util.List;
+import wooteco.subway.domain.Fare;
+import wooteco.subway.domain.Station;
 
 public class PathResult {
     private final int fare;
